refactor(github): pivot check-status coalescer to singleflight-only
Drop the TTL+map+eviction layer from the GetPRCheckStatuses coalescer
and keep only the singleflight wrapper. Check status for a SHA is
mutable (reruns, late-arriving checks, branch-protection adding
required checks), so any TTL window risks converting a now-failing
gate into a passing one. Singleflight-only preserves the
burst-collapsing win — a webhook delivery that fans out to multiple
gate checks for the same commit still makes one upstream GraphQL
round trip — while eliminating the stale-success window entirely.

- Rename CheckStatusCache -> CheckStatusSingleflight and the file
  pair to match; the name now describes what the type does.
- Drop NewClientWithCacheTTL and DefaultCheckStatusCacheTTL; NewClient
  is the only constructor and takes no TTL arg.
- Drop the (repo, sha) -> rows+expiry map, its mutex, opportunistic
  eviction on lookup, and the sweep-on-store path.
- Keep the singleflight wrapper with its leader-cancellation fix:
  context.WithoutCancel + context.WithTimeout(30s) inside DoChan, and
  a select on the result channel vs the caller's ctx.Done() so a
  leader cancellation cannot abort the shared fetch and fail
  unrelated waiters.
- Keep the per-call IsSchemaBot recompute in GetPRCheckStatuses; the
  singleflight delivers one shared result to N waiters and each
  waiter's InstallationClient classifies rows independently against
  its own appSlug snapshot.
- Drop the TTL/eviction tests. Keep concurrent-collapse,
  waiter-respects-own-ctx, leader-cancellation-doesn't-fail-waiters,
  and per-waiter-IsSchemaBot-recompute. Add a
  sequential-calls-refetch test that locks in the "no memoisation
  across calls" invariant the pivot delivers.
